Refactor:#76 - standardize close icon naming, location, and format
- Relocate from public/assets/icons/closeIcon.svg to src/assets/icons/close.svg
- Apply consistent SVGR props
